New module structure, Exner equation, timestep
